requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
python-docx==0.8.11  # For template processing
openai==1.3.7       # For AI processing
pytest              # For testing
